Move TypedEnvelope into legacy package where it's still used and useful
